manifest: move to platform arrays
Rather than specifying platforms by type we now support arbitrary
amounts of platform configurations. Each configuration requires a
unique ID to identify it, and then uses the old platform-type as
detailed configuration.

With this model, we can now have multiple configurations for the same
platform but with different configuration values. This will be required
for platforms that have multiple deployment options. With the new model
we are much more flexible and do not require fixed deployment types.

Signed-off-by: David Rheinsberg <david@readahead.eu>