manifest: provide path converter
Add a path converter to the manifest to actually support relative paths
from the manifest location. This requires keeping track of the path
where we got the manifest from, and then properly converting to absolute
paths when necessary.

Preferably, the type system would support a distinction between both
paths, but yet here we are and have to deal with the limited path
support.

Signed-off-by: David Rheinsberg <david@readahead.eu>